Crate ReportValidator

+++ source/DailyReport.BusinessLogic/Servicies/ReportService.cs
@@ -1,8 +1,10 @@
-﻿using DailyReport.BusinessLogic.Interfaces;
+﻿using DailyReport.BusinessLogic.ExceptionValidators;
+using DailyReport.BusinessLogic.Interfaces;
 using DailyReport.BusinessLogic.Mappers;
 using DailyReport.BusinessLogic.ModelsDTO;
 using DailyReport.DataAccess.Interfaces;
 using DailyReport.DataAccess.Models;
+using static DailyReport.BusinessLogic.Exceptions.ExceptionValidator;
 
 namespace DailyReport.BusinessLogic.Servicies
 {
@@ -19,8 +21,15 @@ namespace DailyReport.BusinessLogic.Servicies
         {
             if (item == null)
             {
-                //
-                throw new ArgumentNullException("item");
+                throw new ValidationException("Can not create a report");
+            }
+            if (ReportValidator.PersonExistsInReport(item.PersonId, _reportRepository))
+            {
+                throw new ValidationException($"This Person already exists in the report");
+            }
+            if (ReportValidator.StartTimeExistsInReport(item.StartTime, _reportRepository))
+            {
+                throw new ValidationException($"Start Time'{item.StartTime}' already exists in the report");
             }
             Report report = ReportMapper.FromDTO(item);
             await _reportRepository.CreateAsync(report);
@@ -28,6 +37,14 @@ namespace DailyReport.BusinessLogic.Servicies
 
         public async Task DeleteAsync(ReportDTO item)
         {
+            if (item == null)
+            {
+                throw new ValidationException("Can not delete a report");
+            }
+            if (item.Id <= 0)
+            {
+                throw new ValidationException("It is impossible");
+            }
             Report report = ReportMapper.FromDTO(item);
             await _reportRepository.DeleteAsync(report);
         }
@@ -46,6 +63,14 @@ namespace DailyReport.BusinessLogic.Servicies
 
         public async Task<ReportDTO> GetByIdAsync(int? id)
         {
+            if (id == null)
+            {
+                throw new ValidationException("Can not get a report");
+            }
+            if (id <= 0)
+            {
+                throw new ValidationException("It is impossible");
+            }
             var report = await _reportRepository.GetByIdAsync(id);
             ReportDTO reportDTO = ReportMapper.ToDTO(report);
             return reportDTO;
@@ -53,6 +78,14 @@ namespace DailyReport.BusinessLogic.Servicies
 
         public async Task UpdateAsync(ReportDTO item)
         {
+            if (item == null)
+            {
+                throw new ValidationException("Can not update a report");
+            }
+            if (item.Id <= 0)
+            {
+                throw new ValidationException("It is impossible");
+            }
             Report report = ReportMapper.FromDTO(item);
             await _reportRepository.UpdateAsync(report);
         }
